+++ Workshop/src/Workshop/UserPolling.java
@@ -4,7 +4,7 @@ public interface UserPolling {
 	public void writeSavedInstructionsInSavefile();
 	public void recordInstructions();
 	public void sendFile();
-	void saveOneInstruction(byte codeOp, String description, int nbArg,String[] descriptionArgs, byte[] args);
+	void saveOneInstruction(short codeOp, String description, int nbArg,String[] descriptionArgs, short[] args);
 	void requestDisplayOfPrimitiveInstructions();
 	
 }
